Added some Moderation Commands
Added the commands: muted, unmute, clear, ban, unban and hello.

Notes: The mute command is not working. The bot gives the "Muted" role to the user, but the user can still talk and speak in voice channels. Also when banning/kicking a member from a server, the bot should send also a private message to that user.